phase2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.2.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
